Ajuste de arquitetura em gerar relatório, para carregar os dados nescessario para a operação de gerar relatório do Animal.

+++ Gestor De Pule/src/Views/Cadastros/Animal/FormAnimalAtualizar.cs
@@ -10,7 +10,7 @@ namespace Gestor_De_Pule.src.Views
             InitializeComponent();
             _controller = new AnimalController();
            //AnimalController.AnimalSelecionado(animalSelecionado);
-            _controller.AnimalSelecionado(animalSelecionado);
+            _controller.AnimalSelecionado((int)animalSelecionado);
             SetAnimalData();
         }
